backend_2/requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0

# Database
databases>=0.5.5
aiosqlite>=0.19.0
sqlalchemy>=2.0.0

# Agent system
crewai>=0.1.0
crewai[tools]>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
